+++ src/main/java/calculator/domain/Number.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.domain;
 
 public class Number {
