+++ src/test/java/calculator/NumberTest.java
@@ -1,5 +1,6 @@
 package calculator;
 
+import calculator.domain.Number;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
