+++ src/test/java/calculator/StringCalculatorTest.java
@@ -1,5 +1,6 @@
 package calculator;
 
+import calculator.utils.StringCalculator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
